requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
requests
polygon-api-client
alpaca-py
alembic